refactor: extract text annotation actions
